refactor local tracks into LocalTracks.kt, clean up playable select serializer
