+++ src/commonMain/kotlin/com.adamratzman.spotify/models/Playable.kt
@@ -2,8 +2,6 @@
 package com.adamratzman.spotify.models
 
 import kotlinx.serialization.KSerializer
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
 import kotlinx.serialization.json.JsonElement
 import kotlinx.serialization.json.JsonObject
 import kotlinx.serialization.json.JsonParametricSerializer
@@ -17,49 +15,14 @@ interface Playable {
 }
 
 object PlayableTransformingSerializer : JsonParametricSerializer<Playable>(Playable::class) {
-    private val spotifyLocalUriRegex = "spotify:local:.+".toRegex()
-
     override fun selectSerializer(element: JsonElement): KSerializer<out Playable> {
-        element as JsonObject
-        return when {
-            element["uri"]?.contentOrNull?.matches(spotifyLocalUriRegex) == true -> LocalTrack.serializer()
+        val uri = (element as? JsonObject)?.get("uri")?.contentOrNull?.let { PlayableUri(it) }
+        return when (uri) {
+            is LocalTrackUri -> LocalTrack.serializer()
             else -> Track.serializer()
         }
     }
 }
-
-@Serializable
-data class SimpleLocalAlbum(
-        @SerialName("album_type") private val albumTypeString: String? = null,
-        val artists: List<SimpleLocalArtist> = listOf(),
-        val name: String,
-        @SerialName("release_date") private val releaseDate: String? = null,
-        @SerialName("release_date_precision") val releaseDatePrecisionString: String? = null,
-        val type: String
-)
-
-@Serializable
-data class SimpleLocalArtist(
-        val name: String,
-        val type: String
-)
-
-@Serializable
-data class LocalTrack(
-        val album: SimpleLocalAlbum,
-        val artists: List<SimpleLocalArtist>,
-        override val href: String? = null,
-        override val id: String? = null,
-        @SerialName("disc_number") val discNumber: String? = null,
-        @SerialName("duration_ms") val durationMs: Int? = null,
-        @SerialName("explicit") val explicit: Boolean? = null,
-        @SerialName("is_local") val isLocal: Boolean = true,
-        val name: String,
-        val popularity: Int? = null,
-        @SerialName("track_number") val trackNumber: Int? = null,
-        override val type: String,
-        override val uri: LocalTrackUri
-) : IdentifiableNullable(), Playable
 
 /*
 @Serializable
