+++ src/commonMain/kotlin/com.adamratzman.spotify/models/SpotifyUris.kt
@@ -7,10 +7,10 @@ import kotlinx.serialization.Encoder
 import kotlinx.serialization.KSerializer
 import kotlinx.serialization.PrimitiveDescriptor
 import kotlinx.serialization.PrimitiveKind
+import kotlinx.serialization.PrimitiveKind.STRING
 import kotlinx.serialization.SerialDescriptor
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.Serializer
-import kotlinx.serialization.internal.StringDescriptor
 
 /**
  * Exception instantiating or deserializing a uri perceived as invalid
@@ -51,7 +51,7 @@ private class SimpleUriSerializer<T : SpotifyUri>(val ctor: (String) -> T) : KSe
  * @property id representation of this uri as an id
  */
 @Serializable
-sealed class SpotifyUri(input: String, type: String, allowColon: Boolean = false) {
+sealed class SpotifyUri(input: String, val type: String, allowColon: Boolean = false) {
     val uri: String
     val id: String
 
@@ -74,7 +74,7 @@ sealed class SpotifyUri(input: String, type: String, allowColon: Boolean = false
     }
 
     override fun toString(): String {
-        return "SpotifyUri($uri)"
+        return "SpotifyUri(type=$type, uri=$uri)"
     }
 
     @Serializer(forClass = SpotifyUri::class)
@@ -181,11 +181,11 @@ sealed class ImmutableCollectionUri(input: String, type: String, allowColon: Boo
 }
 
 @Serializable
-sealed class PlayableUri(input: String, val type: String, allowColon: Boolean = false) :
+sealed class PlayableUri(input: String, type: String, allowColon: Boolean = false) :
     SpotifyUri(input, type, allowColon) {
     @Serializer(forClass = PlayableUri::class)
     companion object : KSerializer<PlayableUri> {
-        override val descriptor: SerialDescriptor = StringDescriptor
+        override val descriptor: SerialDescriptor = PrimitiveDescriptor("PlayableUri", STRING)
         override fun deserialize(decoder: Decoder): PlayableUri = PlayableUri(decoder.decodeString())
         override fun serialize(encoder: Encoder, value: PlayableUri) = encoder.encodeString(value.uri)
