+++ src/commonTest/kotlin/com.adamratzman/spotify/public/PublicPlaylistsAPITest.kt
@@ -3,6 +3,8 @@ package com.adamratzman.spotify.public
 
 import com.adamratzman.spotify.SpotifyException
 import com.adamratzman.spotify.api
+import com.adamratzman.spotify.models.LocalTrack
+import com.adamratzman.spotify.models.Track
 import kotlin.test.assertEquals
 import kotlin.test.assertFailsWith
 import kotlin.test.assertNull
@@ -37,8 +39,10 @@ class PublicPlaylistsAPITest : Spek({
                 assertTrue(p.getPlaylistTracks("78eWnYKwDksmCHAjOUNPEj").complete().items.isNotEmpty())
             }
 
-            it("valid playlist, get tracks including local tracks") {
-                println(p.getPlaylistTracks("0vzdw0N41qZLbRDqyx2cE0").complete().items.map { it.track?.let { it::class } })
+            it("valid playlist, get tracks including one local track") {
+                val playlist = p.getPlaylistTracks("0vzdw0N41qZLbRDqyx2cE0").complete()
+                assertEquals(LocalTrack::class, playlist[0].track!!::class)
+                assertEquals(Track::class, playlist[1].track!!::class)
             }
 
             it("invalid playlist") {
